Implemented Archon as talent by talent and not central analyzer
