+++ src/analysis/retail/priest/holy/modules/talents/Archon/ManifestedPowerHoly.tsx
@@ -1,4 +1,4 @@
-import Analyzer from 'parser/core/Analyzer';
+import Analyzer, { SELECTED_PLAYER } from 'parser/core/Analyzer';
 import { Options } from 'parser/core/Module';
 import Combatants from 'parser/shared/modules/Combatants';
 import Statistic from 'parser/ui/Statistic';
@@ -6,11 +6,23 @@ import STATISTIC_CATEGORY from 'parser/ui/STATISTIC_CATEGORY';
 import STATISTIC_ORDER from 'parser/ui/STATISTIC_ORDER';
 import SpellLink from 'interface/SpellLink';
 import { formatNumber } from 'common/format';
-import ArchonAnalysis from './ArchonAnalysis';
 import BoringSpellValueText from 'parser/ui/BoringSpellValueText';
 
 import { TALENTS_PRIEST } from 'common/TALENTS';
 import PRIEST_TALENTS from 'common/TALENTS/priest';
+import Events, {
+  ApplyBuffEvent,
+  ApplyBuffStackEvent,
+  RefreshBuffEvent,
+  RemoveBuffEvent,
+  RemoveBuffStackEvent,
+} from 'parser/core/Events';
+import SPELLS from 'common/SPELLS';
+import {
+  buffedBySurgeOfLight,
+  getHealFromSurge,
+  isSurgeOfLightFromHalo,
+} from '../../../normalizers/CastLinkNormalizer';
 /**
  * **Perfected Form**
  * Your healing done is increased by 10% while Apotheosis is active and for 20 sec after you cast Holy Word: Salvation.
@@ -20,16 +32,84 @@ import PRIEST_TALENTS from 'common/TALENTS/priest';
 class ManifestedPowerHoly extends Analyzer {
   static dependencies = {
     combatants: Combatants,
-    archonanalysis: ArchonAnalysis,
   };
 
   protected combatants!: Combatants;
-  protected archonanalysis!: ArchonAnalysis;
+
+  /**
+   * Start:
+   * These values can all be called directly and don't need a pass function
+   * also check the call functions in case math needs to be done to transform
+   * the results
+   */
+  surgeOfLightProcsSpent = 0;
+  surgeOfLightProcsGainedTotal = 0;
+  surgeOfLightProcsOverwritten = 0;
+  surgeOfLightProcsOverwrittenByHalo = 0;
+  surgeOfLightProcsGainedFromHalo = 0;
 
   constructor(options: Options) {
     super(options);
 
-    this.active = this.selectedCombatant.hasTalent(TALENTS_PRIEST.MANIFESTED_POWER_TALENT);
+    //if the capstone is active, then all the analyzed hero talents will be active too
+    this.active = this.selectedCombatant.hasTalent(TALENTS_PRIEST.DIVINE_HALO_TALENT);
+
+    //tracks spending of Surge of Light
+    this.addEventListener(
+      Events.removebuff.by(SELECTED_PLAYER).spell(SPELLS.SURGE_OF_LIGHT_BUFF),
+      this.onSurgeOfLightHeal,
+    );
+
+    this.addEventListener(
+      Events.removebuffstack.by(SELECTED_PLAYER).spell(SPELLS.SURGE_OF_LIGHT_BUFF),
+      this.onSurgeOfLightHeal,
+    );
+
+    //tracks gains of Surge of light, not just from manifested power
+    this.addEventListener(
+      Events.applybuff.by(SELECTED_PLAYER).spell(SPELLS.SURGE_OF_LIGHT_BUFF),
+      this.onApplySurgeOfLight,
+    );
+    this.addEventListener(
+      Events.applybuffstack.by(SELECTED_PLAYER).spell(SPELLS.SURGE_OF_LIGHT_BUFF),
+      this.onApplySurgeOfLight,
+    );
+    //tracks wasted Surge of Lights
+    this.addEventListener(
+      Events.refreshbuff.by(SELECTED_PLAYER).spell(SPELLS.SURGE_OF_LIGHT_BUFF),
+      this.onRefreshSurgeOfLight,
+    );
+  }
+
+  onSurgeOfLightHeal(event: RemoveBuffEvent | RemoveBuffStackEvent) {
+    // linked heal event exists from surge of light consumption
+    const healEvent = getHealFromSurge(event);
+
+    if (healEvent) {
+      if (buffedBySurgeOfLight(event)) {
+        this.surgeOfLightProcsSpent += 1;
+      }
+    }
+  }
+
+  onRefreshSurgeOfLight(event: RefreshBuffEvent) {
+    // surge of light used successfully (spending on 2 stacks counts as a refresh)
+    if (1 === this.selectedCombatant.getBuffStacks(SPELLS.SURGE_OF_LIGHT_BUFF.id)) {
+      return;
+    }
+    if (isSurgeOfLightFromHalo(event)) {
+      this.surgeOfLightProcsOverwrittenByHalo += 1;
+    } else {
+      this.surgeOfLightProcsOverwritten += 1;
+    }
+  }
+
+  onApplySurgeOfLight(event: ApplyBuffEvent | ApplyBuffStackEvent) {
+    // linked heal event exists
+    if (isSurgeOfLightFromHalo(event)) {
+      this.surgeOfLightProcsGainedFromHalo += 1;
+    }
+    this.surgeOfLightProcsGainedTotal += 1;
   }
 
   statistic() {
@@ -44,15 +124,15 @@ class ManifestedPowerHoly extends Analyzer {
             sources:
             <ul>
               <li>
-                {formatNumber(this.archonanalysis.surgeOfLightProcsGainedTotal)}
+                {formatNumber(this.surgeOfLightProcsGainedTotal)}
                 {' gained total'}
               </li>
               <li>
-                {formatNumber(this.archonanalysis.surgeOfLightProcsSpent)}
+                {formatNumber(this.surgeOfLightProcsSpent)}
                 {' spent total'}
               </li>
               <li>
-                {formatNumber(this.archonanalysis.surgeOfLightProcsOverwritten)}
+                {formatNumber(this.surgeOfLightProcsOverwritten)}
                 {' overwritten'}
               </li>
             </ul>
@@ -61,9 +141,9 @@ class ManifestedPowerHoly extends Analyzer {
       >
         <BoringSpellValueText spell={TALENTS_PRIEST.MANIFESTED_POWER_TALENT}>
           <div>
-            {this.archonanalysis.surgeOfLightProcsGainedFromHalo}
+            {this.surgeOfLightProcsGainedFromHalo}
             <small> procs gained from Halo</small> <br />
-            {this.archonanalysis.surgeOfLightProcsOverwrittenByHalo}
+            {this.surgeOfLightProcsOverwrittenByHalo}
             <small> overwritten procs from Halo</small> <br />
           </div>
         </BoringSpellValueText>
