+++ src/analysis/retail/priest/holy/modules/talents/Archon/index.tsx
@@ -1,6 +1,5 @@
 import PerfectedFormHoly from './PerfectedFormHoly';
 import ResonantEnergyHoly from './ResonantEnergyHoly';
-import ArchonAnalysis from './ArchonAnalysis';
 import ManifestedPowerHoly from './ManifestedPowerHoly';
 import EmpoweredSurgesHoly from './EmpoweredSurgesHoly';
 import EnergyCycleHoly from './EnergyCycleHoly';
@@ -10,7 +9,6 @@ import PowerSurgeAndDivineHaloHoly from './PowerSurgeAndDivineHaloHoly';
 export {
   PerfectedFormHoly,
   ResonantEnergyHoly,
-  ArchonAnalysis,
   ManifestedPowerHoly,
   EmpoweredSurgesHoly,
   EnergyCycleHoly,
